+++ packages/realm-react/src/__tests__/useQueryHook.test.tsx
@@ -88,12 +88,17 @@ describe("useQuery", () => {
     const [dog1, dog2, dog3] = testDataSet;
 
     expect(collection).not.toBeNull();
-    expect(collection?.length).toBe(6);
+    expect(collection.length).toBe(6);
+    expect(collection[0]).toMatchObject(dog1);
+    expect(collection[1]).toMatchObject(dog2);
+    expect(collection[2]).toMatchObject(dog3);
+  });
+  it("returns the same collection reference if there are no changes", () => {
+    const { result } = renderHook(() => useQuery<IDog>("dog"));
+    const collection = result.current;
 
-    if (collection !== undefined) {
-      expect(collection?.[0]).toMatchObject(dog1);
-      expect(collection?.[1]).toMatchObject(dog2);
-      expect(collection?.[2]).toMatchObject(dog3);
-    }
+    expect(collection).not.toBeNull();
+    expect(collection.length).toBe(6);
+    expect(collection[0]).toEqual(collection?.[0]);
   });
 });