+++ templates/react-native-template-realm-js/template/app/models/Task.js
@@ -1,11 +1,12 @@
-import { Realm, createRealmContext } from '@realm/react';
-
-export class Task {
-  constructor({id = new Realm.BSON.ObjectId(), description, isComplete = false}) {
-    this.description = description;
-    this.isComplete = isComplete;
-    this.createdAt = new Date();
-    this._id = id;
+import {Realm, createRealmContext} from '@realm/react';
+export class Task extends Realm.Object {
+  static generate(description) {
+    return {
+      _id: new Realm.BSON.ObjectId(),
+      description,
+      isComplete: false,
+      createdAt: new Date(),
+    };
   }
 
   // To use a class as a Realm object type, define the object schema on the static property "schema".
@@ -16,12 +17,12 @@ export class Task {
       _id: 'objectId',
       description: 'string',
       isComplete: {type: 'bool', default: false},
-      createdAt: 'date'
+      createdAt: 'date',
     },
   };
 }
 
 export default createRealmContext({
-  schema: [Task.schema],
+  schema: [Task],
   deleteRealmIfMigrationNeeded: true,
 });